+++ cmd/zfs_object_agent/server/Cargo.toml
@@ -17,6 +17,7 @@ nvpair = { git = "https://github.com/ahrens/rust-libzfs", branch = "work"}
 rand = "0.8.3"
 rust-s3 = "0.27.0-rc1"
 serde = { version = "1.0.125", features = ["derive"] }
+stream-reduce = "0.1.0"
 tokio = { version = "1.4", features = ["full"] }
 tokio-pipe = "0.2.1"
 tokio-stream = "0.1.5"
