+++ cmd/zfs_object_agent/server/src/main.rs
@@ -1,4 +1,10 @@
-#[tokio::main]
-async fn main() {
-    libzoa::server::do_server().await;
+fn main() {
+    tokio::runtime::Builder::new_multi_thread()
+        .enable_all()
+        .thread_name("zoa")
+        .build()
+        .unwrap()
+        .block_on(async move {
+            libzoa::server::do_server().await;
+        });
 }
