+++ cmd/zfs_object_agent/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod object_access;
 pub mod object_based_log;
+pub mod object_block_map;
 pub mod pool;
 pub mod server;
